+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* You can now undo and redo changes to loadouts while editing them.
+
 ## 7.39.0 <span class="changelog-date">(2022-10-16)</span>
 
 * Added `is:armorintrinsic` to find Artifice Armor, armor with seasonal perks, etc.